+++ src/main/java/com/vshvet/firstrelease/payload/Request/SaveOrUpdateEngineRequest.java
@@ -12,7 +12,7 @@ public class SaveOrUpdateEngineRequest {
     private Integer cylindersNumber;
     private Integer flapNumber;
     private Double pistonDiameter;
-    private Integer pistonStroke;
+    private Double pistonStroke;
     private int engineCapacity;
     private String powerKwt;
     private Integer horsepower;
@@ -34,7 +34,7 @@ public class SaveOrUpdateEngineRequest {
                                      Integer cylindersNumber,
                                      Integer flapNumber,
                                      Double pistonDiameter,
-                                     Integer pistonStroke,
+                                     Double pistonStroke,
                                      int engineCapacity,
                                      String powerKwt,
                                      Integer horsepower,
@@ -142,11 +142,11 @@ public class SaveOrUpdateEngineRequest {
         this.pistonDiameter = pistonDiameter;
     }
 
-    public Integer getPistonStroke() {
+    public Double getPistonStroke() {
         return pistonStroke;
     }
 
-    public void setPistonStroke(Integer pistonStroke) {
+    public void setPistonStroke(Double pistonStroke) {
         this.pistonStroke = pistonStroke;
     }
