+++ src/main/java/com/vshvet/firstrelease/payload/Response/AutoEngAndParamResponse.java
@@ -16,7 +16,7 @@ public class AutoEngAndParamResponse {
     private String powerKWT;
     private Integer engineCapacity;
     private Integer horsepower;
-    private Integer pistonStoke;
+    private Double pistonStoke;
     private Double pistonDiameter;
     private String modelName;
     private String engineType;
@@ -166,11 +166,11 @@ public class AutoEngAndParamResponse {
         this.pistonDiameter = pistonDiameter;
     }
 
-    public Integer getPistonStoke() {
+    public Double getPistonStoke() {
         return pistonStoke;
     }
 
-    public void setPistonStoke(Integer pistonStoke) {
+    public void setPistonStoke(Double pistonStoke) {
         this.pistonStoke = pistonStoke;
     }
